moved some src into project for x64
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@6 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ networking/ipaddress.cpp
@@ -43,9 +43,9 @@ IPAddress::IPAddress(const std::string& addstring) throw (MalformedIPStringExcep
 
 	int a, b, c, d;
 
-	int nextstartpos;
+	size_t nextstartpos;
 	{
-	const int dotpos = addstring.find_first_of('.');
+	const size_t dotpos = addstring.find_first_of('.');
 	if(dotpos == std::string::npos)
 		throw MalformedIPStringExcep();
 
@@ -54,7 +54,7 @@ IPAddress::IPAddress(const std::string& addstring) throw (MalformedIPStringExcep
 	}
 
 	{
-	const int dotpos = addstring.find_first_of('.', nextstartpos);
+	const size_t dotpos = addstring.find_first_of('.', nextstartpos);
 	if(dotpos == std::string::npos)
 		throw MalformedIPStringExcep();
 
@@ -67,7 +67,7 @@ IPAddress::IPAddress(const std::string& addstring) throw (MalformedIPStringExcep
 	}
 
 	{
-	const int dotpos = addstring.find_first_of('.', nextstartpos);
+	const size_t dotpos = addstring.find_first_of('.', nextstartpos);
 	if(dotpos == std::string::npos)
 		throw MalformedIPStringExcep();
